backend/requirements.txt
asgiref==3.8.1
attrs==25.3.0
Django==5.2
django-cors-headers==4.7.0
django-filter==25.1
django-jalali==7.4.0
django-js-asset==3.1.2
django-mptt==0.17.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
inflection==0.5.1
jalali_core==1.0.0
jdatetime==5.2.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
pillow==11.2.1
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-decouple==3.8
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.24.0
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.1.1
